pyproject.toml
[tool.poetry]
name = "aimai-search"
version = "0.1.0"
description = ""
authors = ["reiyw <reiyw.setuve@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
cutlet = "^0.1.3"
rapidfuzz = "^0.9.1"
attrs = "^19.3.0"
more-itertools = "^8.4.0"
unidic-lite = "^1.0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
